[#IOPID-2838] moved config file
