+++ apps/io-profile/src/config.ts
@@ -19,7 +19,7 @@ import { readableReport } from "@pagopa/ts-commons/lib/reporters";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { withFallback } from "io-ts-types";
 import { UrlFromString } from "@pagopa/ts-commons/lib/url";
-import { FeatureFlag, FeatureFlagEnum } from "./featureFlag";
+import { FeatureFlag, FeatureFlagEnum } from "./utils/featureFlag";
 
 // exclude a specific value from a type
 // as strict equality is performed, allowed input types are constrained to be values not references (object, arrays, etc)
