+++ apps/io-profile/src/utils/unique_email_enforcement.ts
@@ -1,6 +1,6 @@
 import { TableClient } from "@azure/data-tables";
 
-import { getConfigOrThrow } from "./config";
+import { getConfigOrThrow } from "../config";
 
 const config = getConfigOrThrow();
